requirements.txt
Django==2.2.3
django-filter==2.2.0
djangorestframework==3.10.0
Markdown==3.1.1
mysqlclient==1.4.2.post1
Pillow==6.1.0
pytz==2019.1
sqlparse==0.3.0
